requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pgvector==0.3.5
alembic==1.13.2
python-multipart==0.0.9
redis==5.1.0
rq==1.16.2
transformers==4.44.2
torch==2.4.1
torchvision==0.19.1
Pillow==10.4.0
httpx==0.27.2
pydantic==2.9.2
pydantic-settings==2.5.2
